batch_pipeline/requirements.txt
attrs==22.2.0 ; python_version >= "3.9" and python_version < "4.0"
aws-cdk-asset-awscli-v1==2.2.127 ; python_version >= "3.9" and python_version < "4.0"
aws-cdk-asset-kubectl-v20==2.1.1 ; python_version >= "3.9" and python_version < "4.0"
aws-cdk-asset-node-proxy-agent-v5==2.0.104 ; python_version >= "3.9" and python_version < "4.0"
aws-cdk-lib==2.72.1 ; python_version >= "3.9" and python_version < "4.0"
boto3==1.26.104 ; python_version >= "3.9" and python_version < "4.0"
botocore==1.29.104 ; python_version >= "3.9" and python_version < "4.0"
cattrs==22.2.0 ; python_version >= "3.9" and python_version < "4.0"
constructs==10.1.300 ; python_version >= "3.9" and python_version < "4.0"
contextlib2==21.6.0 ; python_version >= "3.9" and python_version < "4.0"
dill==0.3.6 ; python_version >= "3.9" and python_version < "4.0"
exceptiongroup==1.1.1 ; python_version >= "3.9" and python_version < "3.11"
google-pasta==0.2.0 ; python_version >= "3.9" and python_version < "4.0"
importlib-metadata==4.13.0 ; python_version >= "3.9" and python_version < "4.0"
importlib-resources==5.12.0 ; python_version >= "3.9" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.9" and python_version < "4.0"
jsii==1.79.0 ; python_version >= "3.9" and python_version < "4.0"
jsonschema==4.17.3 ; python_version >= "3.9" and python_version < "4.0"
multiprocess==0.70.14 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.24.2 ; python_version < "4.0" and python_version >= "3.9"
packaging==23.0 ; python_version >= "3.9" and python_version < "4.0"
pandas==1.5.3 ; python_version >= "3.9" and python_version < "4.0"
pathos==0.3.0 ; python_version >= "3.9" and python_version < "4.0"
platformdirs==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
pox==0.3.2 ; python_version >= "3.9" and python_version < "4.0"
ppft==1.7.6.6 ; python_version >= "3.9" and python_version < "4.0"
protobuf3-to-dict==0.1.5 ; python_version >= "3.9" and python_version < "4.0"
protobuf==3.20.3 ; python_version >= "3.9" and python_version < "4.0"
publication==0.0.3 ; python_version >= "3.9" and python_version < "4.0"
pyrsistent==0.19.3 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==5.4.1 ; python_version >= "3.9" and python_version < "4.0"
s3transfer==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
sagemaker==2.143.0 ; python_version >= "3.9" and python_version < "4.0"
schema==0.7.5 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
smdebug-rulesconfig==1.0.1 ; python_version >= "3.9" and python_version < "4.0"
typeguard==2.13.3 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.15 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.15.0 ; python_version >= "3.9" and python_version < "4.0"
